Split domain stats metrics to virt-handler and VMI status metrics to virt-controller
