+++ cmd/virt-handler/BUILD.bazel
@@ -12,9 +12,9 @@ go_library(
         "//pkg/healthz:go_default_library",
         "//pkg/inotify-informer:go_default_library",
         "//pkg/monitoring/client/prometheus:go_default_library",
+        "//pkg/monitoring/domainstats/downwardmetrics:go_default_library",
+        "//pkg/monitoring/domainstats/prometheus:go_default_library",
         "//pkg/monitoring/reflector/prometheus:go_default_library",
-        "//pkg/monitoring/vms/downwardmetrics:go_default_library",
-        "//pkg/monitoring/vms/prometheus:go_default_library",
         "//pkg/monitoring/workqueue/prometheus:go_default_library",
         "//pkg/service:go_default_library",
         "//pkg/util:go_default_library",
