+++ pkg/monitoring/domainstats/collector.go
