+++ pkg/monitoring/domainstats/collector_test.go
