+++ pkg/monitoring/domainstats/downwardmetrics/BUILD.bazel
@@ -6,14 +6,14 @@ go_library(
         "hostmetrics.go",
         "scraper.go",
     ],
-    importpath = "kubevirt.io/kubevirt/pkg/monitoring/vms/downwardmetrics",
+    importpath = "kubevirt.io/kubevirt/pkg/monitoring/domainstats/downwardmetrics",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/downwardmetrics:go_default_library",
         "//pkg/downwardmetrics/vhostmd:go_default_library",
         "//pkg/downwardmetrics/vhostmd/api:go_default_library",
         "//pkg/downwardmetrics/vhostmd/metrics:go_default_library",
-        "//pkg/monitoring/vms:go_default_library",
+        "//pkg/monitoring/domainstats:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-launcher/virtwrap/stats:go_default_library",
