+++ pkg/monitoring/domainstats/downwardmetrics/testdata/cpuinfo
