+++ pkg/monitoring/domainstats/prometheus/BUILD.bazel
@@ -6,11 +6,10 @@ go_library(
         "fakeCollector.go",
         "prometheus.go",
     ],
-    importpath = "kubevirt.io/kubevirt/pkg/monitoring/vms/prometheus",
+    importpath = "kubevirt.io/kubevirt/pkg/monitoring/domainstats/prometheus",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/controller:go_default_library",
-        "//pkg/monitoring/vms:go_default_library",
+        "//pkg/monitoring/domainstats:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-launcher/virtwrap/stats:go_default_library",
         "//pkg/virt-launcher/virtwrap/statsconv:go_default_library",
@@ -21,7 +20,6 @@ go_library(
         "//staging/src/kubevirt.io/client-go/version:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus/promhttp:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
         "//vendor/libvirt.org/libvirt-go:go_default_library",
     ],
@@ -39,11 +37,9 @@ go_test(
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
-        "//vendor/github.com/onsi/ginkgo/extensions/table:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/github.com/prometheus/client_model/go:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/libvirt.org/libvirt-go:go_default_library",
     ],
