+++ pkg/monitoring/domainstats/scraper.go
