+++ pkg/monitoring/vmistats/BUILD.bazel
@@ -4,29 +4,35 @@ go_library(
     name = "go_default_library",
     srcs = [
         "collector.go",
-        "scraper.go",
+        "fakecollector.go",
     ],
-    importpath = "kubevirt.io/kubevirt/pkg/monitoring/vms",
+    importpath = "kubevirt.io/kubevirt/pkg/monitoring/vmistats",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/virt-handler/cmd-client:go_default_library",
+        "//pkg/controller:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/client-go/tools/cache:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
     srcs = [
-        "collector_suite_test.go",
         "collector_test.go",
+        "vmistats_suite_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
-        "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
+        "//vendor/github.com/onsi/ginkgo/extensions/table:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
+        "//vendor/github.com/prometheus/client_model/go:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
     ],
 )
