+++ pkg/monitoring/vmistats/vmistats_suite_test.go
@@ -1,0 +1,13 @@
+package vmistats_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestVmistats(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Vmistats Suite")
+}
